+++ src/mesido/financial_mixin.py
@@ -16,6 +16,9 @@ from rtctools.optimization.timeseries import Timeseries
 
 logger = logging.getLogger("mesido")
 
+# No safety factor is needed for big_m, since actual bounds are used.
+ASSET_IS_REALIZED_BIG_M_MARGIN = 1.01
+
 
 class FinancialMixin(
     BaseProblemMixin, BaseComponentTypeMixin, CollocatedIntegratedOptimizationProblem
@@ -1163,9 +1166,15 @@ class FinancialMixin(
             variable_operational_cost_coefficient = parameters[  # euro / Wh
                 f"{es}.variable_operational_cost_coefficient"
             ]
+
+            price_profile = self.__get_electricity_price_profile_or_zero()
+
             sum_ = ca.sum1(
                 variable_operational_cost_coefficient * elec_produced_w[1:] * timesteps_hr
             )  # [euro/Wh] * [W] * [hr]
+            if es in self.energy_system_components.get("electricity_import", []):
+                sum_ += ca.sum1(price_profile[1:] * elec_produced_w[1:] * timesteps_hr)
+
             constraints.append(((variable_operational_cost - sum_) / nominal, 0.0, 0.0))
 
         # for a in self.heat_network_components.get("ates", []):
@@ -1294,6 +1303,7 @@ class FinancialMixin(
         """
         constraints = []
         options = self.energy_system_options()
+        bounds = self.bounds()
         if options["include_asset_is_realized"] and not options["yearly_investments"]:
             for asset in [
                 *self.energy_system_components.get("heat_demand", []),
@@ -1318,10 +1328,10 @@ class FinancialMixin(
                 #         insulation_class_cost
                 #         investment_cost_sym += insulation_class_active * insulation_class_cost
                 big_m = (
-                    1.5
+                    ASSET_IS_REALIZED_BIG_M_MARGIN
                     * max(
-                        self.bounds()[f"{asset}__investment_cost"][1]
-                        + self.bounds()[f"{asset}__installation_cost"][1],
+                        bounds[f"{asset}__investment_cost"][1]
+                        + bounds[f"{asset}__installation_cost"][1],
                         1.0,
                     )
                     / max(self.get_aggregation_count_max(asset), 1.0)
@@ -1350,28 +1360,28 @@ class FinancialMixin(
 
                 # Once the asset is utilized the asset must be realized
                 heat_flow = self.state(f"{asset}.Heat_flow")
-                if not np.isinf(self.bounds()[f"{asset}.Heat_flow"][1]):
+                if not np.isinf(bounds[f"{asset}.Heat_flow"][1]):
                     big_m = (
-                        1.5
-                        * self.bounds()[f"{asset}.Heat_flow"][1]
+                        ASSET_IS_REALIZED_BIG_M_MARGIN
+                        * bounds[f"{asset}.Heat_flow"][1]
                         / max(self.get_aggregation_count_max(asset), 1.0)
                     )
                 else:
                     try:
                         big_m = (
-                            1.5
+                            ASSET_IS_REALIZED_BIG_M_MARGIN
                             * max(
-                                self.bounds()[f"{asset}.HeatOut.Heat"][1],
-                                self.bounds()[f"{asset}.HeatIn.Heat"][1],
+                                bounds[f"{asset}.HeatOut.Heat"][1],
+                                bounds[f"{asset}.HeatIn.Heat"][1],
                             )
                             / max(self.get_aggregation_count_max(asset), 1.0)
                         )
                     except KeyError:
                         big_m = (
-                            1.5
+                            ASSET_IS_REALIZED_BIG_M_MARGIN
                             * max(
-                                self.bounds()[f"{asset}.Primary.HeatOut.Heat"][1],
-                                self.bounds()[f"{asset}.Primary.HeatIn.Heat"][1],
+                                bounds[f"{asset}.Primary.HeatOut.Heat"][1],
+                                bounds[f"{asset}.Primary.HeatIn.Heat"][1],
                             )
                             / max(self.get_aggregation_count_max(asset), 1.0)
                         )
@@ -1397,6 +1407,7 @@ class FinancialMixin(
         """
         constraints = []
         options = self.energy_system_options()
+        bounds = self.bounds()
         if options["include_asset_is_realized"] and options["yearly_investments"]:
             for asset in [
                 *self.energy_system_components.get("heat_demand", []),
@@ -1423,10 +1434,10 @@ class FinancialMixin(
                     )
 
                     big_m = (
-                        1.5
+                        ASSET_IS_REALIZED_BIG_M_MARGIN
                         * max(
-                            self.bounds()[f"{asset}__investment_cost"][1]
-                            + self.bounds()[f"{asset}__installation_cost"][1],
+                            bounds[f"{asset}__investment_cost"][1]
+                            + bounds[f"{asset}__installation_cost"][1],
                             1.0,
                         )
                         / max(self.get_aggregation_count_max(asset), 1.0)
@@ -1450,28 +1461,28 @@ class FinancialMixin(
 
                     # Once the asset is utilized the asset must be realized
                     heat_flow = self.states_in(f"{asset}.Heat_flow", time_start, time_end)[:-1]
-                    if not np.isinf(self.bounds()[f"{asset}.Heat_flow"][1]):
+                    if not np.isinf(bounds[f"{asset}.Heat_flow"][1]):
                         big_m = (
-                            1.5
-                            * self.bounds()[f"{asset}.Heat_flow"][1]
+                            ASSET_IS_REALIZED_BIG_M_MARGIN
+                            * bounds[f"{asset}.Heat_flow"][1]
                             / max(self.get_aggregation_count_max(asset), 1.0)
                         )
                     else:
                         try:
                             big_m = (
-                                1.5
+                                ASSET_IS_REALIZED_BIG_M_MARGIN
                                 * max(
-                                    self.bounds()[f"{asset}.HeatOut.Heat"][1],
-                                    self.bounds()[f"{asset}.HeatIn.Heat"][1],
+                                    bounds[f"{asset}.HeatOut.Heat"][1],
+                                    bounds[f"{asset}.HeatIn.Heat"][1],
                                 )
                                 / max(self.get_aggregation_count_max(asset), 1.0)
                             )
                         except KeyError:
                             big_m = (
-                                1.5
+                                ASSET_IS_REALIZED_BIG_M_MARGIN
                                 * max(
-                                    self.bounds()[f"{asset}.Primary.HeatOut.Heat"][1],
-                                    self.bounds()[f"{asset}.Primary.HeatIn.Heat"][1],
+                                    bounds[f"{asset}.Primary.HeatOut.Heat"][1],
+                                    bounds[f"{asset}.Primary.HeatIn.Heat"][1],
                                 )
                                 / max(self.get_aggregation_count_max(asset), 1.0)
                             )