+++ CHANGELOG.md
@@ -1,10 +1,29 @@
+
 # [Unreleased-main] - 2026-08-24
 
 ## Added
-- xxx
+- Import with an Electricity port is now supported as a subclass of electricity source and caters for electricity price profiles.
 
 ## Changed
 - Use pipe DN enums instead of extracting the DN size from the name (EDR pipes and available pipe classes)
+- Write result profiles to a fixed (configured) database combined with a dynamic schema name per run when the output profile type is postgres
+- Update pyESDL to v26.7.1
+- Parsing of input profiles with 15-minute interval now supported.
+- Class name of ModelicaComponentTypeMixin has been changed to ComponentTypeMixin to reflect the fact that it is not only used for Modelica components.
+
+## Fixed
+- Bug: minimum velocity is set to 0.0 in the staged grow_workflow to allow for zero heat demand.
+
+
+# [0.1.20.3] - 2026-07-15
+
+## Added
+- ESDL profile reading and writing for different types: influxdb and postgres, and inline (in ESDL file).
+
+## Changed
+- Write_result_db_profiles has been replaced by esdl_profiles_output_type (determines the ESDL output profile type)
+- Names used for keys in "database_connections" specified
+- Using pyesdl for database credential management
 
 ## Fixed
 - xxx 
@@ -391,5 +410,4 @@
  
 ## Fixed
 - Error exit when stage 1, priority 1 was not successful
-